refactor(db): update PoolHub DbContext relationships and constraints

+++ PoolHub.Core/DTOs/Order/OrderDtos.cs
@@ -1,5 +1,5 @@
 namespace PoolHub.Core.DTOs.Order;
 
-public class OrderDto { public int OrderId { get; set; } public int SessionId { get; set; } public int Status { get; set; } }
-public class CreateOrderRequest { public int SessionId { get; set; } }
-public class AddOrderItemRequest { public int ProductId { get; set; } public int Quantity { get; set; } }
+public class OrderDto { public long OrderId { get; set; } public long SessionId { get; set; } public int Status { get; set; } }
+public class CreateOrderRequest { public long SessionId { get; set; } }
+public class AddOrderItemRequest { public long ProductId { get; set; } public int Quantity { get; set; } }
