+++ PoolHub.Core/Interfaces/CrudInterfaces.cs
@@ -9,28 +9,28 @@ namespace PoolHub.Core.Interfaces;
 public interface ICrudService
 {
     Task<PagedResult<FloorDto>> GetFloorsAsync(PaginationRequest request, CancellationToken ct);
-    Task<FloorDto> GetFloorAsync(int id, CancellationToken ct);
+    Task<FloorDto> GetFloorAsync(long id, CancellationToken ct);
     Task<FloorDto> CreateFloorAsync(FloorDto dto, CancellationToken ct);
-    Task<FloorDto> UpdateFloorAsync(int id, FloorDto dto, CancellationToken ct);
-    Task DeleteFloorAsync(int id, CancellationToken ct);
+    Task<FloorDto> UpdateFloorAsync(long id, FloorDto dto, CancellationToken ct);
+    Task DeleteFloorAsync(long id, CancellationToken ct);
 
     Task<PagedResult<ZoneDto>> GetZonesAsync(PaginationRequest request, CancellationToken ct);
-    Task<ZoneDto> GetZoneAsync(int id, CancellationToken ct);
+    Task<ZoneDto> GetZoneAsync(long id, CancellationToken ct);
     Task<ZoneDto> CreateZoneAsync(ZoneDto dto, CancellationToken ct);
-    Task<ZoneDto> UpdateZoneAsync(int id, ZoneDto dto, CancellationToken ct);
-    Task DeleteZoneAsync(int id, CancellationToken ct);
+    Task<ZoneDto> UpdateZoneAsync(long id, ZoneDto dto, CancellationToken ct);
+    Task DeleteZoneAsync(long id, CancellationToken ct);
 
     Task<PagedResult<TableTypeDto>> GetTableTypesAsync(PaginationRequest request, CancellationToken ct);
-    Task<TableTypeDto> GetTableTypeAsync(int id, CancellationToken ct);
+    Task<TableTypeDto> GetTableTypeAsync(long id, CancellationToken ct);
     Task<TableTypeDto> CreateTableTypeAsync(TableTypeDto dto, CancellationToken ct);
-    Task<TableTypeDto> UpdateTableTypeAsync(int id, TableTypeDto dto, CancellationToken ct);
-    Task DeleteTableTypeAsync(int id, CancellationToken ct);
+    Task<TableTypeDto> UpdateTableTypeAsync(long id, TableTypeDto dto, CancellationToken ct);
+    Task DeleteTableTypeAsync(long id, CancellationToken ct);
 
     Task<PagedResult<VenueTableDto>> GetVenueTablesAsync(PaginationRequest request, CancellationToken ct);
-    Task<VenueTableDto> GetVenueTableAsync(int id, CancellationToken ct);
+    Task<VenueTableDto> GetVenueTableAsync(long id, CancellationToken ct);
     Task<VenueTableDto> CreateVenueTableAsync(VenueTableDto dto, CancellationToken ct);
-    Task<VenueTableDto> UpdateVenueTableAsync(int id, VenueTableDto dto, CancellationToken ct);
-    Task DeleteVenueTableAsync(int id, CancellationToken ct);
+    Task<VenueTableDto> UpdateVenueTableAsync(long id, VenueTableDto dto, CancellationToken ct);
+    Task DeleteVenueTableAsync(long id, CancellationToken ct);
 
     Task<PagedResult<PricingPlanDto>> GetPricingPlansAsync(PaginationRequest request, CancellationToken ct);
     Task<PagedResult<PricingPlanRuleDto>> GetPricingPlanRulesAsync(PaginationRequest request, CancellationToken ct);
@@ -39,6 +39,6 @@ public interface ICrudService
     Task<ProductCategoryDto> CreateProductCategoryAsync(ProductCategoryDto dto, CancellationToken ct);
 
     Task<PagedResult<BookingDto>> GetBookingsCrudAsync(PaginationRequest request, CancellationToken ct);
-    Task<BookingDto> GetBookingAsync(int id, CancellationToken ct);
-    Task DeleteBookingAsync(int id, CancellationToken ct);
+    Task<BookingDto> GetBookingAsync(long id, CancellationToken ct);
+    Task DeleteBookingAsync(long id, CancellationToken ct);
 }
