+++ PoolHub.Shared/Extensions/ClaimsPrincipalExtensions.cs
@@ -4,9 +4,9 @@ namespace PoolHub.Shared.Extensions;
 
 public static class ClaimsPrincipalExtensions
 {
-    public static int GetUserId(this ClaimsPrincipal user)
+    public static long GetUserId(this ClaimsPrincipal user)
     {
         var raw = user.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? user.FindFirst("sub")?.Value;
-        return int.TryParse(raw, out var id) ? id : 0;
+        return long.TryParse(raw, out var id) ? id : 0;
     }
 }
